lambda_build/requirements.txt
fastapi==0.115.6
uvicorn==0.30.6
python-multipart==0.0.12
pydantic==2.9.2
pydantic-settings==2.5.2

requests==2.32.3
httpx==0.27.2

SQLAlchemy==2.0.36
psycopg2-binary==2.9.9

rapidfuzz==3.10.0
numpy==2.1.2
jinja2==3.1.4
markdown==3.7

openai==1.59.3
pymupdf==1.25.1
weasyprint==64.0
python-docx==1.1.0
beautifulsoup4==4.12.3
lxml==5.1.0
bcrypt==4.0.1
mangum

